Update ome-faq.yml
Added a third limitation to the known issues with decryption and shared mailboxes. After a VERY lengthy discussion with MSFT support, we have been informed that this limitation is correct. If you have a access to a shared mailbox with automapping and can view the decrypted emails in OWA without issues they won't decrypt in the Outlook Desktop client. This is an extension of the Outlook Desktop client not supporting delegated access which is the same as Shared mailbox full access.